# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Patronage20 API",
    "description": "Backend service for smart-home application.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [],
  "components": {
    "schemas": {
      "DashboardGet200": {
        "$ref": "schemas/dashboard-get-200.json"
      },
      "GenericError": {
        "$ref": "schemas/generic-error.json"
      },
      "MapAddRequestBody":{
        "$ref": "schemas/map-add-request-body.json"
      },
      "SensorId": {
        "$ref": "schemas/sensor-id.json"
      },
      "NotificationsGet200": {
        "$ref": "schemas/notifications-get-200.json"
      }
    },
    "responses": {
      "400": {
        "description": "Request malformed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "schemas/generic-error.json"
            }
          }
        }
      },
      "404": {
        "description": "Entity not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "schemas/generic-error.json"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/.well-known/documentation": {
      "get": {
        "summary": "This documentation",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "summary": "Endpoint with all data needed for dashboard visualisation.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "schemas/dashboard-get-200.json"
                },
                "examples": {
                  "positive": {
                    "$ref": "examples/positive.json#/dashboard-get-200"
                  },
                  "positive_example": {
                    "$ref": "examples/positive_example.json#/dashboard-get-200"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "schemas/generic-error.json"
                }
              }
            }
          }
        }
      }
    },

    "/api/v1/map/{id}": {
      "post": {
        "summary": "Endpoint to add sensor to map",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "schemas/map-add-request-body.json"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Endpoint to delete sensor from map",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
            }
          }
        }
      },
      "responses": {
        "200": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "schemas/map-add-request-body.json"
              },
              "examples": {
                "positive": {
                  "$ref": "examples/positive.json#/api-v1-map-{sensorID}-add"
                },
                "positive_example": {
                  "$ref": "examples/positive_example.json#/api-v1-map-{sensorID}-add"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "summary": "Endpoint with data containing information about notifications.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "schemas/notifications-get-200.json"
                },
                "examples": {
                  "positive": {
                    "$ref": "examples/positive/notifications-get-200.json"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "schemas/generic-error.json"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{id}": {
      "delete": {
        "summary": "Endpoint to check notifications as seen",
        "requestBody": {
          "required": false
        }
      }
    }
  }
}